LEVDEN-3021 -  Define Admin API (#1)
* Define API contract

* 3021- Remove Countries classes and and JPA Enum annotation

* 3021- Change tables structure

* 3021- Change tables structure

* 3021- Add new entity model

* 3021 - Add Foreign Key to CompanyEntity model

* 3021- Change endpoint for creation company code

* Change from Access code to Company code

* 3021 - Add cascade relationship between tables

* 3021- add custom constraint to CodeType

* 3021 - Check admin authorization

* 3021 - Add logging messages

* 3021 - Add one trial test

* 3021- Add Controller Exception Handler

* Define API contract

* 3021- Remove Countries classes and and JPA Enum annotation

* 3021- Change tables structure

* 3021- Change tables structure

* 3021- Add new entity model

* 3021 - Add Foreign Key to CompanyEntity model

* 3021- Change endpoint for creation company code

* Change from Access code to Company code

* 3021 - Add cascade relationship between tables

* 3021- add custom constraint to CodeType

* 3021 - Check admin authorization

* 3021 - Add logging messages

* 3021 - Add one trial test

* Define API contract

* 3021- Remove Countries classes and and JPA Enum annotation

* 3021- Change tables structure

* 3021- Change tables structure

* 3021- Add new entity model

* 3021 - Add Foreign Key to CompanyEntity model

* 3021- Change endpoint for creation company code

* Change from Access code to Company code

* 3021 - Add cascade relationship between tables

* 3021- add custom constraint to CodeType

* 3021 - Check admin authorization

* 3021 - Add logging messages

* 3021- Add Controller Exception Handler

* added example of unit test which saves into H2 DB

* Add swagger

* 3021 - Add unit tests for code and authorization service

* 3021 - Add one trial test

* added example of unit test which saves into H2 DB

* 3021 - Add unit tests for code and authorization service

* 3021 - Rebase to feature branch

* 3021 - Add new exception handler method for CodeController

* added SecurityContext, request filter to populate it with access code and cleaned up Services signature methods

* fixed failing test

* edited OpenAPIDefinition description of CodesController

---------

+++ ppac-service-impl/src/main/kotlin/com/levi9/ppac/service/api/service/CompanyServiceImpl.kt
@@ -1,37 +1,42 @@
 package com.levi9.ppac.service.api.service
 
 import com.levi9.ppac.service.api.data_classes.Company
+import com.levi9.ppac.service.api.repository.CodeRepository
 import com.levi9.ppac.service.api.repository.CompanyRepository
+import com.levi9.ppac.service.api.security.SecurityContext
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import org.springframework.web.server.ResponseStatusException
 import java.util.*
 
 @Service
 @ConditionalOnProperty(prefix = "feature", name = ["mvp"], havingValue = "true")
 class CompanyServiceImpl(
-    val companyRepository: CompanyRepository
-): CompanyService<Company> {
+    private val securityContext: SecurityContext<Int>,
+    private val companyRepository: CompanyRepository,
+    private val codeRepository: CodeRepository
+) : CompanyService<Company> {
     @Transactional
     override fun findAll(): List<Company> {
+
+        if (!codeRepository.isAdminCode(securityContext.getAccessCode())) {
+            throw ResponseStatusException(HttpStatus.UNAUTHORIZED)
+        }
+
         return companyRepository.findAll().map { Company.parse(it) }
     }
 
     @Transactional
-    override fun create(dto: Company): Company {
-        val persistedCompany = companyRepository.save(
-            Company.parse(dto).apply { id = UUID.randomUUID() }
-        )
-        return Company.parse(persistedCompany)
-    }
-
-    @Transactional
     override fun deleteById(id: UUID) {
+
+        if (!codeRepository.isAdminCode(securityContext.getAccessCode())) {
+            throw ResponseStatusException(HttpStatus.UNAUTHORIZED)
+        }
+
         companyRepository.findByIdOrNull(id)?.let {
-            it.students?.forEach { student ->
-                student.companyId = null
-            }
             companyRepository.deleteById(id)
         }
     }
